Work-in-progress with custom activation function (tanh and ReLU)
This should result in for faster training as demonstrated in the test code. However at the moment, the loss is quite good (<10e-3) but results are gibberish, probably due to some miscoding of the new normalisation procedures. Also note that, test code is creating its own training code (forward pass and back-propagation), which is better probably. So the old training function and the new normalisation functions inside MLP are commented out for the time being.